requirements.txt
fastapi==0.115.4
tensorflow==2.18.0
python-multipart==0.0.17
pillow==11.0.0
tensorflow==2.18.0
numpy==2.0.2
pydantic==2.9.2
SQLAlchemy==2.0.36
mysql==0.0.3
pymysql==1.1.1
cryptography==43.0.3
python-dotenv==1.0.1